Fixed Articles status count delay issue
